+++ services/identity/src/main/java/com/crapi/service/UserRegistrationService.java
@@ -1,0 +1,26 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the “License”);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an “AS IS” BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.crapi.service;
+
+import com.crapi.entity.User;
+import com.crapi.entity.UserDetails;
+import com.crapi.model.*;
+
+public interface UserRegistrationService {
+
+  CRAPIResponse registerUser(SignUpForm signUpRequest);
+
+  UserDetails createUserDetails(String name, User user);
+}
