+++ services/workshop/build-image.sh
@@ -17,3 +17,8 @@
 set -x
 cd "$(dirname $0)"
 docker build -t crapi/crapi-workshop:${VERSION:-latest} .
+retVal=$?
+if [ $retVal -ne 0 ]; then
+    echo "Error building image"
+fi
+exit $retVal
